Add resize for inputoverlay-key; Exit some for loops when files are already deleted

+++ 0.imouji_pack.sh
@@ -26,6 +26,7 @@ done
 if ls *.png 2>/dev/null >/dev/null; then
   echo trimming totrim images
   for i in *totrim.png; do
+    [ ! -f $i ] && continue
     convert -trim +repage $i $(basename $i totrim.png).png
     rm $i
   done
@@ -49,14 +50,23 @@ done
 for f in 8 4; do
   echo resizing @"$f"x to @$((f/2))x and removing @"$f"x...
   for i in *@"$f"x.png; do
-    convert -resize 50% $i $(basename $i @"$f"x.png)@$((f/2))x.png >/dev/null 2>/dev/null # just silence it
-    rm $i >/dev/null 2>/dev/null
+    [ ! -f $i ] && continue
+    convert -resize 50% $i $(basename $i @"$f"x.png)@$((f/2))x.png
+    rm $i
   done
 done
 
 echo resizing score-dot and score-comma...
 for i in score-{dot,comma}@2xtmp.png; do
+  [ ! -f $i ] && continue
   convert -crop 20x84+14+0 $i $(basename $i @2xtmp.png)@2x.png
+  rm $i
+done
+
+echo resizing inputoverlay-key...
+for i in inputoverlay-key@2xtmp.png; do
+  [ ! -f $i ] && continue
+  convert -resize 86x86 $i $(basename $i @2xtmp.png)@2x.png
   rm $i
 done
